refactor(test): consolidate Kamino test files into solana-kamino.test.ts
Mirror solana-marginfi.test.ts structure (one file per Solana DeFi
protocol). Merges solana-kamino-actions.test.ts (init_user + supply
builders) into solana-kamino-pr3-pr4.test.ts (borrow/withdraw/repay
+ reader + render-verification), then renames to solana-kamino.test.ts.

Setup overlap was ~95% — the Solana RPC mock, nonce mock, kit/sysvars
shims, fakeMarket, VanillaObligation/KaminoObligation/KaminoAction
class mocks, and beforeEach reset block were near-identical across
both files. The only material delta was actions-file needing granular
mocks for buildDepositTxns / initObligation / getUserLutAddressAndSetupIxs
that pr3-pr4 had as inline arrows; merged file uses the granular shape
with defaults that match the inline-arrow behavior so the borrow/
withdraw/repay tests still see what they expect.

1116 tests / 92 files (was 93). Coverage preserved.
